[refactor] delegate 분리, model 인스턴스 위치 변경 #24
