+++ iOS/ToDo/ToDo/View/TaskView/DataSource/TaskDataSource.swift
@@ -8,9 +8,14 @@
 
 import UIKit
 
-class TaslDataSource: NSObject, UITableViewDataSource {
+class TaskDataSource: NSObject, UITableViewDataSource {
     
-    var model: Category?
+    var model: Category? {
+        didSet {
+            handler()
+        }
+    }
+    var handler: () -> () = {}
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return model?.count ?? 0
